+++ packages/web3-utils/test/unit/socket_provider.test.ts
@@ -312,7 +312,7 @@ describe('SocketProvider', () => {
 				});
 				// @ts-expect-error run protected method
 				expect(provider._pendingRequestsQueue.size).toBe(1);
-
+				expect(provider.getPendingRequestQueueSize()).toBe(1);
 				const payload2 = { id: 2, method: 'some_rpc_method' };
 				provider.setStatus('connected');
 				const req2 = provider.request(payload2);
@@ -331,6 +331,7 @@ describe('SocketProvider', () => {
 				provider._clearQueues();
 				// @ts-expect-error run protected method
 				expect(provider._pendingRequestsQueue.size).toBe(0);
+				expect(provider.getPendingRequestQueueSize()).toBe(0);
 				// @ts-expect-error run protected method
 				expect(provider._sentRequestsQueue.size).toBe(0);
 			});